+++ src/value.rs
@@ -966,8 +966,7 @@ pub trait ReprValue: private::ReprValue {
             R: BlockReturn,
         {
             let func = std::mem::transmute::<VALUE, fn(&[Value], Option<Proc>) -> R>(callback_arg);
-            Block::new(func)
-                .call_handle_error(argc, argv as *const Value, Value::new(blockarg))
+            func.call_handle_error(argc, argv as *const Value, Value::new(blockarg))
                 .as_rb_value()
         }